requirements.txt
pytest==7.4.3
pytest-html==3.2.0
requests==2.31.0
urllib3==2.0.7
selenium==4.11.2
webdriver-manager==3.8.6
PyYAML==6.0.1
regex==2023.10.3
pytest-cov==4.1.0
pytest-xdist==3.3.1
python-dotenv==1.0.0
allure-pytest==2.13.2
watchdog==3.0.0 